% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_pol.R
\name{analyse_pol}
\alias{analyse_pol}
\title{Analyze Primary-Secondary Data}
\usage{
analyse_pol(df, reference = "Ctrl", na.rm = TRUE)
}
\arguments{
\item{df}{data.frame: should have columns 'primary', 'secondary', and 'response'}

\item{reference}{string: Name of the reference level within the 'primary'
variable/column. Default is "Ctrl"}

\item{na.rm}{logical: indicates how mean() and sd() will treat NAs downstream}
}
\value{
tibble with columns primary, secondary, mean, sd, and p.value.
}
\description{
Takes a data.frame with observations from multiple primary-secondary
conditions and calculates the mean and standard deviation for each
primary-secondary group. Also performs a t.test on each group relative
to the reference primary group with the same secondary stimulation.
Result is suitable for use with plot_pol().
}
\examples{
df <- tibble::tibble(
  'primary' = rep(rep(c("Ctrl", "M1"), each=3), 2),
  'secondary'= rep(c("Ctrl", "LPS"), each=6),
  'response' = c(1, 2, 3, 4, 5, 6, 4, NA, 6, 11, 12, 13)
)
analyse_pol(df)
analyse_pol(df, na.rm = FALSE)

df2 <- tibble::tibble(
  'primary' = rep(rep(c("WT", "Mutant"), each=3), 2),
  'secondary'= rep(c("Diet1", "Diet2"), each=6),
  'response' = c(1, 2, 3, 4, 5, 6, 4, 5, 6, 11, 12, 13)
)
analyse_pol(df2, reference = "WT")
}

