% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pstar.R
\name{get_pstar}
\alias{get_pstar}
\title{Translate p-value to Star Symbols}
\usage{
get_pstar(pval)
}
\arguments{
\item{pval}{double, or vector of doubles: between 0 and 1}
}
\value{
factor: "" "*" "**" or "***"
}
\description{
Translates a numeric p-value into a significance symbol according to
0.05 < p          : ""
0.01 <= p < 0.05  : "*"
0.001 <= p < 0.05 : "**"
p < 0.001         : "***"
}
\examples{
get_pstar(0.8)
get_pstar(0.05)
get_pstar(0.00002)
get_pstar(c(0.8, 0.05, 0.0002))
}

