% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pol.R
\name{plot_pol}
\alias{plot_pol}
\title{Plot Bar Graph with Statistical Summaries}
\usage{
plot_pol(df, xlab = "Polarization", ylab = "Response", label = NULL)
}
\arguments{
\item{df}{data.frame: Requires columns "primary", "secondary", "mean",
and "sd", as well as an optional p.value or p.star column if labels used.}

\item{xlab}{string: x axis label, name of the type of observation of the
primary variable (eg mutation)}

\item{ylab}{string: y axis label, name of response measured}

\item{label}{string or NULL: p.value and p.star values are currently supported.
If only a p.value can be found, p.star can be derived within the function. If
an unsupported label type is given, no label will be used and a warning message
will be produced.}
}
\value{
plot
}
\description{
Produces a bar graph with the primary variable being represented as facets
and the secondary variable represented by position along the x-axis and colour.
Group mean is plotted, along with error bars representing the standard deviation
and an optional label indicating statistical significance (numeric p-value
or stars, see get_pstar). Axis labels may be customised to represent
the specific data used.
}
\examples{
df <- tibble::tibble(
  'primary' = c("Ctrl", "M1", "Ctrl", "M1"),
  'secondary'= c("Ctrl", "Ctrl", "M1", "M1"),
  'mean' = c(2, 5, 5, 12),
  'sd' = c(1, 1, 1.414, 1),
  'p.value' = c(1, 0.21, 1, 0.38)
)
plot_pol(df)
plot_pol(df, ylab="Inflammation", label="p.value")
plot_pol(df, ylab="Inflammation", label="p.star")
plot_pol(df, ylab="Inflammation", label="sd") #produces warning, no label
}

