% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pol.R
\name{tidy_pol}
\alias{tidy_pol}
\title{Tidy Primary-Secondary Data}
\usage{
tidy_pol(data)
}
\arguments{
\item{data}{data.frame where each column represents a treatment group,
the first row contains treatment names in form primary-secondary
(eg 'Ctrl-LPS'), and subsequent rows contain individual observations.}
}
\value{
tibble with columns primary <chr>, secondary <chr> and
response <int> or <dbl>
}
\description{
Takes a data.frame or tibble in a specific human-friendly format
and tidies it into the conventional format suitable for use with
analyse_pol()
}
\examples{
data <- tibble::tibble(
  'Ctrl-Ctrl' = 1:3,
  'M1-Ctrl'= 4:6,
  'Ctrl-LPS' = 4:6,
  'M1-LPS' = 11:13
)
tidy_pol(data)
}

