% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pstar.R
\name{get_pstar}
\alias{get_pstar}
\title{Translate p-value to Star Symbols}
\usage{
get_pstar(pval)
}
\arguments{
\item{pval}{double: between 0 and 1. Note function is not currently vectorized.}
}
\value{
factor: "" "*" "**" or "***"
}
\description{
Translates a numeric p-value into a significance symbol according to
p > 0.05          : ""
0.05 <= p < 0.01  : "*"
0.01 <= p < 0.001 : "**"
0.001 <= p        : "***"
}
\examples{
get_pstar(0.8)
get_pstar(0.05)
get_pstar(0.00002)
get_pstar(1.2) #produces error
get_pstar(c(0.8, 0.05, 0.0002)) #produces warning, uses 0.8
}

