% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pol.R
\name{tidy_pol}
\alias{tidy_pol}
\title{Tidy Polarization Data}
\usage{
tidy_pol(data)
}
\arguments{
\item{data}{data frame}
}
\value{
tibble
}
\description{
Takes data frame where each column represents a treatment group
with the first row being treatment names in form
polarization-secondary (eg 'Ctrl-LPS') and subsequent rows
containing individual observations. Returns tidied tibble
with columns polarization <chr>, secondary <chr> and
response <int>
}
\examples{
data <- tibble::tibble(
'Ctrl-Ctrl' = 1:3,
'M1-Ctrl'= 1:3,
'Ctrl-LPS' = 4:6,
'M1-LPS' = 11:13
)
tidy_pol(data)
}

